requirements.txt
absl-py==2.3.1
antlr4-python3-runtime==4.9.3
astunparse==1.6.3
certifi==2025.7.14
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.59.0
fsspec==2025.7.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.73.1
h5py==3.14.0
hf-xet==1.1.10
hub-sdk==0.0.24
huggingface-hub==0.35.3
idna==3.10
Jinja2==3.1.6
jmetalpy==1.7.0
keras==3.10.0
kiwisolver==1.4.8
libclang==18.1.1
lightning-utilities==0.15.2
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
ml_dtypes==0.5.1
mockito==1.5.4
mpmath==1.3.0
namex==0.1.0
narwhals==1.48.0
networkx==3.4.2
numpy==2.1.3
omegaconf==2.3.0
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.16.0
pandas==2.3.1
patsy==1.0.1
pillow==11.3.0
plotly==6.2.0
polars==1.34.0
polars-runtime-32==1.34.0
protobuf==5.29.5
py-cpuinfo==9.0.0
PyHamcrest==2.1.0
pyparsing==3.2.3
pytz==2025.2
PyYAML==6.0.3
requests==2.32.4
rich==14.0.0
safetensors==0.6.2
scipy==1.15.3
seaborn==0.13.2
statsmodels==0.14.5
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
thop==0.1.1.post2209072238
timm==0.8.13.dev0
torch==2.6
torch-fidelity==0.3.0
torchmetrics==1.8.2
torchsummary==1.5.1
torchvision==0.21.0
tqdm==4.67.1
tzdata==2025.2
ultralytics==8.3.213
ultralytics-thop==2.0.17
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.17.2
yt-dlp==2025.9.26
